perf: dropbear ssh connect too slow
Both `Dropbear` and `OpenSSH` have the same action when a client connects to a
Server: they reverse query the client address to the DNS Server configured on
the device.
If the IP address of the DNS server configured on the device is faulty or the
speed is slow, there will be a delay in the connection between the client and
the client.
While `OpenSSH` can disable this action by config `UseDNS No` in `sshd_config`,
`Dropbear` has no configurable option.

But we can defconfig `BR2_PACKAGE_DROPBEAR_DISABLE_REVERSEDNS` to
disable reverse DNS lookups on connection. This can be handy on systems without
working DNS, as connections otherwise stall until DNS times out.

reference:
https://blog.csdn.net/pengwupeng2008/article/details/119424022
https://blog.csdn.net/force_eagle/article/details/5896496