feat: support pam dynamic update motd
This motd package seeks to make the /etc/motd (Message of the Day) more dynamic
and valuable, by providing a simple, clean framework for defining scripts whose
output will regularly be written to /etc/motd.
In most Linux distributions, you can directly modify the`/etc/motd` file to
customize any desired prompt message by pasting the prompt message file that
needs to be printed into the file (some executable commands or scripts are only
treated as regular characters/text in the file)/ The text message within
`etc/motd` is fixed and unchanged unless we manually modify it. Therefore, the
customized message in`/etc/motd` is a static MOTD.
In Ubuntu, a set of scripts is provided in the directory`/etc/update motd.d/`.
When the user logs in, they are executed in the order of the number (00-99)
prefix of the script name, and the output of these scripts is saved to the
file`/run/motd.dynamic`. After the end user successfully logs in, it is printed
out on the login screen interface.

When executing some programs in Linux, the user who started them must be
authenticated before executing them, and only after meeting certain requirements
can they be allowed to execute, such as login, su, etc. The identity or state
verification program in Linux is carried out by PAM, which can dynamically load
verification modules. Because the verification content can be dynamically
changed as needed, it can greatly improve the flexibility of verification.
Linux-PAM (Linux Pluggable Authentication Module) is a shared library that
allows local system administrators to freely choose the authentication method
for programs. In other words, without recompiling an application that includes
PAM functionality, the authentication mechanism it uses can be changed. In this
way, even if the local authentication mechanism is upgraded, the program does
not need to be modified.
PAM uses the configuration file`/etc/pam.d/`to manage the authentication method
for the program. The application calls the corresponding configuration file to
call the local authentication module. The module is placed in`/lib/security` to
load dynamic libraries. When we use the su command, the system will prompt you
to enter the password of the root user. This is how the su command is
implemented by calling the PAM module.

Therefore, we will build our own dynamic update motd by imitating the features
and file layout in the Ubuntu distribution, using linux-pam.

TODO:
The dynamic update motd message currently only support serial console, ssh not
support.
So expect the message currently printed support on all logins (console, ssh,
etc).